css cleanup + components cleanup and rename of atomic design components (uppercase first letter)
