--- src/components/molecules/card/index.js
@@ -1,2 +1,0 @@
-import * as React from 'react'
-import classnames from 'classnames/bind'
