+++ src/components/pages/Homepage.js
@@ -6,15 +6,19 @@ import React, { Component } from "react";
 // standard components
 
 // organisms
-import Intro from "../organisms/intro";
-import Section from "../organisms/section";
-import Footer from "../organisms/footer";
+import Intro from "../organisms/Intro";
+import Section from "../organisms/Section";
+import Footer from "../organisms/Footer";
 
-//Replace with CMS!
+// Replace with CMS!
 import bg1 from "./bg1.jpg";
 import bg2 from "./bg2.jpg";
 import logo_dark from "./dark.png";
 import logo_light from "./light.png";
+import icon1 from "./icon1.gif";
+import icon2 from "./icon2.gif";
+import icon3 from "./icon3.gif";
+import section1_img from "./section1.jpg";
 /* Later to be replaced with headless CMS content */
 let navitems = [
   {href:"/why-us", text:"Warum TheUniqueStory?", active:false, type:"text"},
@@ -35,6 +39,27 @@ let companyinfo = [
 ]
 let logos = [
   {light:logo_light,dark:logo_dark}
+]
+// Section 0
+let section0_contents = [
+  {heading:"Drei Gründe zu schöner Haut.<br>Garantiert"},
+  {icon:icon1,text:"Dein Pflegeprogramm wird individuell nur für dich und für deine Haut entwickelt"},
+  {icon:icon2,text:"Experten kümmern sich um deine ganz persönlichen Anforderungen"},
+  {icon:icon3,text:"Deine Haut wird auf Dauer mit ehricher, frischer und natürlicher Pflege made in Austria versorgt"},
+  {btntext:"Beautyprogramm starten",btnhref:"/start"}
+]
+// Section 1
+// Number of paragraphs is flexible. Just add more or less and the view will render accordingly
+let section1_contents = [
+  {heading:"Individuelles Programm"},
+  {img:section1_img},
+  {btntext:"Beautyprogramm starten",btnhref:"/start"},
+  {lead:"Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet."},
+  {paragraph:"Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet."},
+  {paragraph:"Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet."},
+  {paragraph:"Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet."},
+  {paragraph:"Test"},
+  {paragraph:"Number of paragraphs is flexible"},
 ]
 // Rendering of all active organisms
 class Homepage extends Component {
